+++ implementation/sipamato/src/main/java/ch/difty/sipamato/persistance/jooq/stepsetter/PaperUpdateSetStepSetter.java
@@ -19,7 +19,7 @@ public class PaperUpdateSetStepSetter implements UpdateSetStepSetter<PaperRecord
     public UpdateSetMoreStep<PaperRecord> setFieldsFor(UpdateSetFirstStep<PaperRecord> step, Paper e) {
         Asserts.notNull(step, "step");
         Asserts.notNull(e, "entity");
-        Integer id = e.getId();
+        Long id = e.getId();
         Asserts.notNull(id, "entity.id");
         // @formatter:off
         return step
