+++ implementation/sipamato/src/main/java/ch/difty/sipamato/persistance/repository/PaperRepository.java
@@ -2,6 +2,6 @@ package ch.difty.sipamato.persistance.repository;
 
 import ch.difty.sipamato.entity.Paper;
 
-public interface PaperRepository extends GenericRepository<Paper, Integer> {
+public interface PaperRepository extends GenericRepository<Paper, Long> {
 
 }
