+++ implementation/sipamato/src/test/java/ch/difty/sipamato/persistance/jooq/stepsetter/PaperUpdateSetStepSetterTest.java
@@ -62,7 +62,7 @@ public class PaperUpdateSetStepSetterTest extends UpdateSetStepSetterTest<PaperR
 
     @Override
     protected void entityFixture() {
-        when(entityMock.getId()).thenReturn(ID.intValue());
+        when(entityMock.getId()).thenReturn(ID);
         PaperRecordMapperTest.entityFixtureWithoutIdFields(entityMock);
     }
